Changes for issue 1
