+++ README.md
@@ -1,3 +1,3 @@
 # Medium articles
 
-1. [How to Write User-friendly Command Line Interfaces in Python](./how_to_write_user_freindly_cli)
+1. [How to Convert Pure Python Functions to User-friendly CLI](how_to_convert_python_functions/article.md)
