+++ how_to_convert_python_functions/code/cli
@@ -7,7 +7,7 @@ Sample QR Generator
 import os
 import sys
 
-cwd = os.path.dirname(__file__)
+cwd = os.path.dirname(os.path.realpath(__file__))
 
 __version__ = '1.0'
