+++ how_to_convert_python_functions/code/qr_code/__init__.py
@@ -1,0 +1,3 @@
+"""
+All QR code related commands are here
+"""
