+++ how_to_convert_python_functions/code/qr_code/generate.py
@@ -1,0 +1,56 @@
+import shutil
+from datetime import datetime
+from enum import Enum
+from typing import TypedDict
+
+import requests
+
+
+class Manufacturer(Enum):
+    pfizer = 1
+    moderna = 2
+    astrazeneca = 3
+    janssen = 4
+    sinovac = 5
+
+
+class Vaccination(TypedDict):
+    manufacturer: Manufacturer
+    date: datetime
+
+
+class QRCode(TypedDict):
+    name: str
+    birth: str
+    vaccination: list[Vaccination]
+
+
+def generate(name: str, birth: str, *manufacturers: Manufacturer, **dates: str) -> None:
+    """
+    Generate your vaccination QR code.
+
+    Args:
+        name (str): name of the vaccinated person
+        birth (str): birthday of the vaccinated person in YYYY-MM-DD format
+        *manufacturers (Enum): manufacturer of the vaccine
+        **dates (str): date of the vaccination
+
+    Return: None
+    """
+    qr_code = QRCode(
+        name=name,
+        birth=birth,
+        vaccination=[
+            Vaccination(manufacturer=manufacturer, date=date)
+            for manufacturer, date in zip(manufacturers, dates.values())
+        ],
+    )
+    res = requests.get(
+        f"http://api.qrserver.com/v1/create-qr-code/?data={qr_code}", stream=True
+    )
+    if res.status_code != 200:
+        raise Exception("QR code cannot be generated by QR Code Generator API.")
+    with open("qr_code.png", "wb") as f:
+        res.raw.decode_content = True
+        shutil.copyfileobj(res.raw, f)
+    print("QR code has been generated.")
